requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.0.12
Django==4.0.3
django-environ==0.8.1
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
idna==3.4
psycopg2==2.8.4
PyJWT==2.6.0
pytz==2022.7.1
requests==2.27.1
responses==0.20.0
sqlparse==0.4.3
urllib3==1.26.15
